# api/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Petstore - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_
    
    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:8080
...
paths:
...
  /timetable/{timetableTaskID}/notification-params:
    get:
      summary: Get timetable task notification params
      tags:
        - Timetable
        - NotificationParams
      security:
        - bearerAuth: []
      description: Get timetable task notification params
      operationId: getTimetableTaskNotificationParams
      parameters:
        - in: path
          name: timetableTaskID
          schema:
            type: integer
          required: true
          description: Timetable task id
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationParams'
        '405':
          description: Invalid input

    put:
      summary: Set timetable task notification params
      tags:
        - Timetable
        - NotificationParams
      security:
        - bearerAuth: []
      description: Set timetable task notification params
      operationId: setTimetableTaskNotificationParams
      parameters:
        - in: path
          name: timetableTaskID
          schema:
            type: integer
          required: true
          description: Timetable task id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationParams'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationParams'
        '405':
          description: Invalid input
  /notification-params:
    get:
      summary: Set default notification params
      tags:
        - NotificationParams
      security:
        - bearerAuth: []
      description: get default notification params
      operationId: getDefaultNotificationParams
      responses:
        '200':
          description: Successfull operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationParams'
        '405':
          description: Invalid input

    put:
      summary: Set default notification params
      tags:
        - NotificationParams
      security:
        - bearerAuth: []
      description: set default notification params
      operationId: setDefaultNotificationParams
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationParams'
        required: true
      responses:
        '201':
          description: Successfull creation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationParams'
        '405':
          description: Invalid input
components:
  schemas:
...
    NotificationParams:
      type: object
      required:
        - period
        - info
      properties:
        period:
          type: integer
          description: Required time for task in minutes
          minimum: 1
          example: 5
        info:
          $ref: '#/components/schemas/NotificationInfo'
    NotificationInfo:
      type: object
      properties:
        cmd:
          type: boolean
        telegram:
          type: integer
        webhook:
          type: string
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      name: api_key
      in: header